[Add] :: 위젯 정보가 없다면 스킵하는 기능 추가
