--- feature/widget/src/main/java/com/onmi/widget/util/MealTime.kt
@@ -1,10 +1,0 @@
-package com.onmi.widget.util
-
-sealed class MealTime {
-
-    data object Morning : MealTime()
-
-    data object Lunch : MealTime()
-
-    data object Dinner : MealTime()
-}
